+++ app/code/community/Ffuenf/ChildProductCategories/Model/Observer/AssignCategories.php
@@ -18,14 +18,46 @@
 
 class Ffuenf_ChildProductCategories_Model_Observer_AssignCategories extends Varien_Event_Observer
 {
+    /**
+     * Assign the categories of a configurable product to its child products on massaction
+     *
+     * @return Ffuenf_ChildProductCategories_Model_Observer_AssignCategories
+     */
+    public function detectProductAttributeChanges($observer)
+    {
+        $attributesData = $observer->getEvent()->getAttributesData();
+        $productIds     = $observer->getEvent()->getProductIds();
+        $user  = Mage::getSingleton('admin/session')->getUser();
+        foreach ($productIds as $id) {
+            $product = Mage::getModel('catalog/product')->load($id);
+            if ($product->getTypeId() == 'configurable') {
+                $childProducts = Mage::getModel('catalog/product_type_configurable')->getUsedProducts(null, $product);
+                $parentCategoryIds = $product->getCategoryIds();
+                $childProductsIds = array();
+                foreach ($childProducts as $childProduct) {
+                    $childCategoryIds = $parentCategoryIds;
+                    $childProductsIds[] = $childProduct->getId();
+                    $childProduct->setCategoryIds($childCategoryIds);
+                    $childProduct->setIsChanged(true);
+                    $childProduct->save();
+                }
+                Mage::log('The categories of "'.$product->getName().'" ID:'.$product->getId().' has been assign to its child products '.implode(",",$childProductsIds));
+            }
+        }
+        return $this;
+    }
 
     /**
-     * Assign the categories of a configurable product to its child products
+     * Assign the categories of a configurable product to its child products on product save
      *
-     * @return boolean|null
+     * @return Ffuenf_ChildProductCategories_Model_Observer_AssignCategories
      */
-    public function assignCategories($observer)
+    public function detectProductChanges($observer)
     {
+        /**
+        * @var $product Mage_Catalog_Model_Product
+        * @var $user    Mage_Admin_Model_User
+        */
         if (Mage::helper('ffuenf_childproductcategories')->isExtensionActive()) {
             $product = $observer->getEvent()->getProduct();
             if ($product->getTypeId() == 'configurable') {
@@ -40,8 +72,8 @@ class Ffuenf_ChildProductCategories_Model_Observer_AssignCategories extends Vari
                     $childProduct->save();
                 }
                 Mage::log('The categories of "'.$product->getName().'" ID:'.$product->getId().' has been assign to its child products '.implode(",",$childProductsIds));
-                return true;
             }
         }
+        return $this;
     }
 }