+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file is used to list changes made in each version of Ffuenf_ChildProductCategories.
 
+## 1.1.0 (October 29, 2015)
+
+* add configurable timeframe for shell script
+* add datepicker for timeframe
+* add test whether extension uses the old-style admin routing
+* fix massaction logic
+
 ## 1.0.0 (October 26, 2015)
 
 * initial public release